added DungeonMovement class, did not refactor all movement out of GameEngine.
