+++ src/main/java/main/DungeonMovement.java
@@ -1,0 +1,19 @@
+package main;
+
+import tiles.TileType;
+
+public class DungeonMovement {
+
+	public boolean isPassable(TileType tile) {
+		switch(tile) {
+			case TREASURE:
+			case PASSABLE:
+			case PLAYER:
+				return true;
+				
+			default:
+				return false;
+		}
+	}
+	
+}
