+++ src/test/java/tiles/TileTypeTest.java
@@ -31,16 +31,4 @@ public class TileTypeTest {
 			assertEquals(exception.getMessage(), TileType.INVALID_CHARACTER_PROVIDED_MESSAGE + "Z");
 		}
 	}
-	
-	@Test
-	public void passable_types() {
-		assertThat(TileType.isPassable(TileType.TREASURE), equalTo(true));
-		assertThat(TileType.isPassable(TileType.PASSABLE), equalTo(true));
-		assertThat(TileType.isPassable(TileType.PLAYER), equalTo(true));
-	}
-	
-	@Test
-	public void not_passable_types() {
-		assertThat(TileType.isPassable(TileType.NOT_PASSABLE), equalTo(false));
-	}
 }
